module2/audio1/requirements.txt
openai
assemblyai
streamlit
python-dotenv